backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
langchain
langchain-core
langgraph
groq
langchain_groq